dataflow-types: Add protobuf for AvailableCollections
